+++ deployments/terraform/kubeconfig.tf
@@ -7,7 +7,7 @@ data "google_container_cluster" "main" {
 # From https://github.com/terraform-providers/terraform-provider-kubernetes/blob/master/kubernetes/test-infra/gke/main.tf
 
 resource "local_file" "kubeconfig" {
-  filename = "${path.cwd}/.kubeconfig"
+  filename = ".kubeconfig"
 
   content = <<EOF
 apiVersion: v1