+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Specification/Application/CriteriaEvaluation/EvaluatePendingCriteriaSpec.php
@@ -13,11 +13,13 @@ declare(strict_types=1);
 
 namespace Specification\Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation;
 
+use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\EvaluateSpelling;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\Text\EvaluateTitleFormatting;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Enrichment\EvaluateCompletenessOfRequiredAttributes;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluationRegistry;
 use Akeneo\Pim\Automation\DataQualityInsights\Application\EvaluateCriterionInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Model\Write;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\GetPendingCriteriaEvaluationsByProductIdsQueryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Repository\CriterionEvaluationRepositoryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\CriterionCode;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\CriterionEvaluationId;
@@ -26,69 +28,86 @@ use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductId;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Psr\Log\LoggerInterface;
+use Webmozart\Assert\Assert;
 
 class EvaluatePendingCriteriaSpec extends ObjectBehavior
 {
-    public function let(CriterionEvaluationRepositoryInterface $repository, CriteriaEvaluationRegistry $registry, LoggerInterface $logger)
-    {
-        $this->beConstructedWith($repository, $registry, $logger);
+    public function let(
+        CriterionEvaluationRepositoryInterface $repository,
+        CriteriaEvaluationRegistry $registry,
+        GetPendingCriteriaEvaluationsByProductIdsQueryInterface $getPendingCriteriaEvaluationsQuery,
+        LoggerInterface $logger
+    ) {
+        $this->beConstructedWith($repository, $registry, $getPendingCriteriaEvaluationsQuery, $logger);
     }
 
     public function it_evaluates_criteria_for_a_set_of_products(
         CriterionEvaluationRepositoryInterface $repository,
         CriteriaEvaluationRegistry $registry,
-        EvaluateCriterionInterface $evaluateCriterion
+        GetPendingCriteriaEvaluationsByProductIdsQueryInterface $getPendingCriteriaEvaluationsQuery,
+        EvaluateCriterionInterface $evaluateSpelling,
+        EvaluateCriterionInterface $evaluateCompleteness
     ) {
-        $criterionCode = new CriterionCode('completeness');
+        $criterionSpelling = new CriterionCode(EvaluateSpelling::CRITERION_CODE);
+        $criterionCompleteness = new CriterionCode(EvaluateCompletenessOfRequiredAttributes::CRITERION_CODE);
 
-        $criterionA = new Write\CriterionEvaluation(
-            new CriterionEvaluationId('95f124de-45cd-495e-ac58-349086ad6cd4'),
-            $criterionCode,
-            new ProductId(42),
-            new \DateTimeImmutable('2019-10-28 10:41:56'),
-            CriterionEvaluationStatus::pending()
-        );
+        $criteria = [
+            'product_42_spelling' => new Write\CriterionEvaluation(
+                new CriterionEvaluationId('95f124de-45cd-495e-ac58-349086ad6cd4'),
+                $criterionSpelling,
+                new ProductId(42),
+                new \DateTimeImmutable('2019-10-28 10:41:56'),
+                CriterionEvaluationStatus::pending()
+            ),
+            'product_42_completeness' => new Write\CriterionEvaluation(
+                new CriterionEvaluationId('77cf74c8-9254-45f8-9993-5c5bb4e34782'),
+                $criterionCompleteness,
+                new ProductId(42),
+                new \DateTimeImmutable('2019-10-28 10:41:56'),
+                CriterionEvaluationStatus::pending()
+            ),
+            'product_123_spelling' => new Write\CriterionEvaluation(
+                new CriterionEvaluationId('d7bcae1e-30c9-4626-9c4f-d06cae03e77e'),
+                $criterionSpelling,
+                new ProductId(42),
+                new \DateTimeImmutable('2019-10-28 10:41:56'),
+                CriterionEvaluationStatus::pending()
+            )
+        ];
 
-        $criterionB = new Write\CriterionEvaluation(
-            new CriterionEvaluationId('d7bcae1e-30c9-4626-9c4f-d06cae03e77e'),
-            $criterionCode,
-            new ProductId(123),
-            new \DateTimeImmutable('2019-10-28 10:41:56'),
-            CriterionEvaluationStatus::pending()
-        );
+        $criteriaProduct42 = (new Write\CriterionEvaluationCollection())
+            ->add($criteria['product_42_spelling'])
+            ->add($criteria['product_42_completeness']);
+        $criteriaProduct123 = (new Write\CriterionEvaluationCollection())
+            ->add($criteria['product_123_spelling']);
 
-        $repository->findPendingByProductIds([42, 123])->willreturn([$criterionA, $criterionB]);
+        $getPendingCriteriaEvaluationsQuery->execute([42, 123])->willreturn([
+            42 => $criteriaProduct42,
+            123 => $criteriaProduct123
+        ]);
 
-        $registry->get($criterionCode)->willReturn($evaluateCriterion);
-        $evaluateCriterion->evaluate($criterionA)->willReturn(new Write\CriterionEvaluationResult());
-        $evaluateCriterion->evaluate($criterionB)->willReturn(new Write\CriterionEvaluationResult());
+        $registry->get($criterionSpelling)->willReturn($evaluateSpelling);
+        $registry->get($criterionCompleteness)->willReturn($evaluateCompleteness);
+        $evaluateSpelling->evaluate($criteria['product_42_spelling'])->willReturn(new Write\CriterionEvaluationResult());
+        $evaluateSpelling->evaluate($criteria['product_123_spelling'])->willReturn(new Write\CriterionEvaluationResult());
+        $evaluateCompleteness->evaluate($criteria['product_42_completeness'])->willReturn(new Write\CriterionEvaluationResult());
 
-        $repository->update(Argument::that(function ($criterion) use ($criterionA) {
-            return
-                $criterionA->getId() === $criterion->getId() && $criterion->getStatus()->isInProgress();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionA) {
-            return
-                $criterionA->getId() === $criterion->getId() && $criterion->getStatus()->isDone();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionB) {
-            return
-                $criterionB->getId() === $criterion->getId() && $criterion->getStatus()->isInProgress();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionB) {
-            return
-                $criterionB->getId() === $criterion->getId() && $criterion->getStatus()->isDone();
-        }))->shouldBeCalled();
+        $repository->update(Argument::any())->shouldBeCalledTimes(2);
 
         $this->evaluateAllCriteria([42, 123]);
+
+        foreach ($criteria as $criterionEvaluation) {
+            Assert::eq(CriterionEvaluationStatus::done(), $criterionEvaluation->getStatus());
+        }
     }
 
     public function it_continues_to_evaluate_if_an_evaluation_failed(
         CriterionEvaluationRepositoryInterface $repository,
         CriteriaEvaluationRegistry $registry,
+        GetPendingCriteriaEvaluationsByProductIdsQueryInterface $getPendingCriteriaEvaluationsQuery,
         EvaluateCriterionInterface $evaluateCriterion
     ) {
-        $criterionCode = new CriterionCode('completeness');
+        $criterionCode = new CriterionCode(EvaluateSpelling::CRITERION_CODE);
 
         $criterionA = new Write\CriterionEvaluation(
             new CriterionEvaluationId('95f124de-45cd-495e-ac58-349086ad6cd4'),
@@ -106,71 +125,78 @@ class EvaluatePendingCriteriaSpec extends ObjectBehavior
             CriterionEvaluationStatus::pending()
         );
 
-        $repository->findPendingByProductIds([42, 123])->willreturn([$criterionA, $criterionB]);
+        $getPendingCriteriaEvaluationsQuery->execute([42, 123])->willreturn([
+            42 => (new Write\CriterionEvaluationCollection())->add($criterionA),
+            123 => (new Write\CriterionEvaluationCollection())->add($criterionB),
+        ]);
 
         $registry->get($criterionCode)->willReturn($evaluateCriterion);
         $evaluateCriterion->evaluate($criterionA)->willThrow(new \Exception('Evaluation failed'));
         $evaluateCriterion->evaluate($criterionB)->willReturn(new Write\CriterionEvaluationResult());
 
-        $repository->update(Argument::that(function ($criterion) use ($criterionA) {
-            return
-                $criterionA->getId() === $criterion->getId() && $criterion->getStatus()->isInProgress();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionA) {
-            return
-                $criterionA->getId() === $criterion->getId() && $criterion->getStatus()->isError();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionB) {
-            return
-                $criterionB->getId() === $criterion->getId() && $criterion->getStatus()->isInProgress();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionB) {
-            return
-                $criterionB->getId() === $criterion->getId() && $criterion->getStatus()->isDone();
-        }))->shouldBeCalled();
+        $repository->update(Argument::any())->shouldBeCalledTimes(2);
 
         $this->evaluateAllCriteria([42, 123]);
+
+        Assert::eq(CriterionEvaluationStatus::error(), $criterionA->getStatus());
+        Assert::eq(CriterionEvaluationStatus::done(), $criterionB->getStatus());
     }
 
     public function it_evaluates_synchronous_criteria_for_a_set_of_products(
         CriterionEvaluationRepositoryInterface $repository,
         CriteriaEvaluationRegistry $registry,
-        EvaluateCriterionInterface $evaluateCriterion
+        GetPendingCriteriaEvaluationsByProductIdsQueryInterface $getPendingCriteriaEvaluationsQuery,
+        EvaluateCriterionInterface $evaluateSpelling,
+        EvaluateCriterionInterface $evaluateCompleteness
     ) {
-        $criterionACode = new CriterionCode(EvaluateCompletenessOfRequiredAttributes::CRITERION_CODE);
-        $criterionBCode = new CriterionCode(EvaluateTitleFormatting::CRITERION_CODE);
+        $criterionSpelling = new CriterionCode(EvaluateSpelling::CRITERION_CODE);
+        $criterionTitleFormatting = new CriterionCode(EvaluateTitleFormatting::CRITERION_CODE);
 
-        $criterionA = new Write\CriterionEvaluation(
-            new CriterionEvaluationId('95f124de-45cd-495e-ac58-349086ad6cd4'),
-            $criterionACode,
-            new ProductId(42),
-            new \DateTimeImmutable('2019-10-28 10:41:56'),
-            CriterionEvaluationStatus::pending()
-        );
+        $criteria = [
+            'product_42_spelling' => new Write\CriterionEvaluation(
+                new CriterionEvaluationId('95f124de-45cd-495e-ac58-349086ad6cd4'),
+                $criterionSpelling,
+                new ProductId(42),
+                new \DateTimeImmutable('2019-10-28 10:41:56'),
+                CriterionEvaluationStatus::pending()
+            ),
+            'product_42_title' => new Write\CriterionEvaluation(
+                new CriterionEvaluationId('77cf74c8-9254-45f8-9993-5c5bb4e34782'),
+                $criterionTitleFormatting,
+                new ProductId(42),
+                new \DateTimeImmutable('2019-10-28 10:41:56'),
+                CriterionEvaluationStatus::pending()
+            ),
+            'product_123_spelling' => new Write\CriterionEvaluation(
+                new CriterionEvaluationId('d7bcae1e-30c9-4626-9c4f-d06cae03e77e'),
+                $criterionSpelling,
+                new ProductId(42),
+                new \DateTimeImmutable('2019-10-28 10:41:56'),
+                CriterionEvaluationStatus::pending()
+            )
+        ];
 
-        $criterionB = new Write\CriterionEvaluation(
-            new CriterionEvaluationId('d7bcae1e-30c9-4626-9c4f-d06cae03e77e'),
-            $criterionBCode,
-            new ProductId(123),
-            new \DateTimeImmutable('2019-10-28 10:41:56'),
-            CriterionEvaluationStatus::pending()
+        $getPendingCriteriaEvaluationsQuery->execute([42, 123])->willreturn([
+            42 => (new Write\CriterionEvaluationCollection())
+                ->add($criteria['product_42_spelling'])
+                ->add($criteria['product_42_title']),
+            123 => (new Write\CriterionEvaluationCollection())
+                ->add($criteria['product_123_spelling']),
+        ]);
 
-        );
+        $registry->get($criterionTitleFormatting)->shouldNotBeCalled();
+        $evaluateCompleteness->evaluate($criteria['product_42_title'])->shouldNotBeCalled();
 
-        $repository->findPendingByProductIds([42, 123])->willreturn([$criterionA, $criterionB]);
+        $registry->get($criterionSpelling)->willReturn($evaluateSpelling);
+        $evaluateSpelling->evaluate($criteria['product_42_spelling'])->willReturn(new Write\CriterionEvaluationResult());
+        $evaluateSpelling->evaluate($criteria['product_123_spelling'])->willReturn(new Write\CriterionEvaluationResult());
 
-        $registry->get($criterionACode)->willReturn($evaluateCriterion);
-        $evaluateCriterion->evaluate($criterionA)->willReturn(new Write\CriterionEvaluationResult());
-
-        $repository->update(Argument::that(function ($criterion) use ($criterionA) {
-            return
-                $criterionA->getId() === $criterion->getId() && $criterion->getStatus()->isInProgress();
-        }))->shouldBeCalled();
-        $repository->update(Argument::that(function ($criterion) use ($criterionA) {
-            return
-                $criterionA->getId() === $criterion->getId() && $criterion->getStatus()->isDone();
-        }))->shouldBeCalled();
+        $repository->update(Argument::any())->shouldBeCalledTimes(2);
 
         $this->evaluateSynchronousCriteria([42, 123]);
+
+        Assert::eq(CriterionEvaluationStatus::pending(), $criteria['product_42_title']->getStatus());
+        Assert::eq(CriterionEvaluationStatus::done(), $criteria['product_42_spelling']->getStatus());
+        Assert::eq(CriterionEvaluationStatus::done(), $criteria['product_123_spelling']->getStatus());
     }
 }