+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Improvements
 
 - DAPI-697: Data quality - As Julia, when I'm on the DQI page, I want to click the attributes that need improvements and land on the PEF.
+- DAPI-806: Improve criteria evaluations performance
 
 ## New features